setting up view page #4
